Move code a bit more and update public documentation
