+++ s/root.go
@@ -15,6 +15,10 @@ import (
 	"github.com/capatazlib/go-capataz/internal/c"
 )
 
+// rootSupervisorName is the name the root supervisor has, this is used to
+// compare the process current name to the rootSupervisorName
+var rootSupervisorName = ""
+
 // notifyTerminationFn is a callback that gets called when a supervisor is
 // terminating (with or without an error).
 type notifyTerminationFn = func(terminateError)
